push

+++ Modele/Manager.cs
@@ -178,11 +178,13 @@ namespace Modele
             }
             else
             {
-                if (items[0].Id == "999:1" && items[1].Id == "999:2")
+                if(items.Count > 1)
                 {
-                    items.RemoveRange(0, 2);
+                    if (items[0].Id == "999:1" && items[1].Id == "999:2")
+                    {
+                        items.RemoveRange(0, 2);
+                    }
                 }
-
             }
         }
